cleaned file tree
