--- Assets/Damage.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 172f8bda61a2d174196e6ae30fd40344
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
