--- Assets/RageSystem.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 34b0c02898355454484256c2d3b29ee7
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
