--- Assets/Scripts/BetterHorzontal.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 9b2891e83020e9e4e87d99f2a819fb75
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
