+++ Assets/Scripts/HealthSystem-Adapted.cs
