+++ Assets/Scripts/HealthSystem-Adapted.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 84315ba4dab8e6f478bab740c91b4ccb
+guid: 7efc63172ae396a4cace9e50defb4209
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
