+++ Assets/Scripts/HideShow.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 18542cf519f1fde45836f13687f4be62
+guid: 80c7dace6e20b3b4bad74fa10d7b1b5e
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
