+++ Assets/Scripts/Inventory.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 0af737a70628e234b8df9f3369a957e2
+guid: 4ed789c8e5b49514cb840d0825411bbe
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
