--- Assets/Scripts/Movement.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 37d2d43dd97540241bd2bc0cadcc963f
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
