--- Assets/Scripts/Test.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: bf5545b4c904b874e9ac2f3489057fbe
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
